[ADD] Add Dynamic Training & Convert Q-table
Add a Dynamic table to improve training (which required a generalization of the Q agent integration inside of the QueueSimulation class).

Implement a short algorithm to naively convert the Q table to a priority list in an attempt to reach the optimal policy in cases where reward/penalty are not dependent on time.